fix: Move "react-test-renderer" to devDependencies
