+++ package.json
@@ -37,7 +37,6 @@
     "react-range-slider-input": "^3.0.7",
     "react-responsive": "^9.0.2",
     "react-spinners": "^0.11.0",
-    "react-test-renderer": "^17.0.2",
     "react-uid": "^2.3.3",
     "styled-components": "^5.3.11",
     "youtube-player": "^5.6.0",
@@ -90,6 +89,7 @@
     "prettier": "^2.8.8",
     "react-scripts": "4.0.3",
     "react-styleguidist": "^11.1.7",
+    "react-test-renderer": "^17.0.2",
     "semantic-release": "^17.4.6",
     "start-server-and-test": "^2.0.0"
   }
