+++ src/components/Atoms/Button/Button.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "jest-styled-components";
-import renderWithTheme from "../../../hoc/shallowWithTheme";
+import renderWithTheme from "../../../../tests/hoc/shallowWithTheme";
 import Button from "./Button";
 
 it("renders a standard styled link correctly", () => {
