+++ src/components/Atoms/ErrorText/ErrorText.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import 'jest-styled-components';
 
-import renderWithTheme from '../../../hoc/shallowWithTheme';
+import renderWithTheme from '../../../../tests/hoc/shallowWithTheme';
 import ErrorText from './ErrorText';
 
 it('renders correctly', () => {
