+++ src/components/Atoms/Link/Link.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "jest-styled-components";
-import renderWithTheme from "../../../hoc/shallowWithTheme";
+import renderWithTheme from "../../../../tests/hoc/shallowWithTheme";
 import Link from "./Link";
 import { Internal } from "../Icons/index";
