+++ src/components/Atoms/Logo/Logo.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "jest-styled-components";
-import renderWithTheme from "../../../hoc/shallowWithTheme";
+import renderWithTheme from "../../../../tests/hoc/shallowWithTheme";
 import Logo from "./Logo";
 
 it("renders correctly", () => {
