+++ src/components/Atoms/RadioButton/RadioButton.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import 'jest-styled-components';
-import renderWithTheme from '../../../hoc/shallowWithTheme';
+import renderWithTheme from '../../../../tests/hoc/shallowWithTheme';
 import RadioButton from './RadioButton';
 
 it('renders correctly', () => {
