+++ src/components/Atoms/SocialIcons/SocialIcons.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import 'jest-styled-components';
-import renderWithTheme from '../../../hoc/shallowWithTheme';
+import renderWithTheme from '../../../../tests/hoc/shallowWithTheme';
 import SocialIcons from './SocialIcons';
 
 it('renders correctly with Comic Relief links', () => {
