+++ src/components/Molecules/Accordion/Accordion.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import 'jest-styled-components';
-import renderWithTheme from '../../../hoc/shallowWithTheme';
+import renderWithTheme from '../../../../tests/hoc/shallowWithTheme';
 import Accordion from './Accordion';
 import Text from '../../Atoms/Text/Text';
