+++ src/components/Molecules/CardDs/CardDs.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import 'jest-styled-components';
-import renderWithTheme from '../../../hoc/shallowWithTheme';
+import renderWithTheme from '../../../../tests/hoc/shallowWithTheme';
 import CardDs from './CardDs';
 import { defaultData }  from '../../../styleguide/data/data';import { Internal } from '../../Atoms/Icons/index';
