+++ src/components/Molecules/Chip/Chip.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "jest-styled-components";
-import renderWithTheme from "../../../hoc/shallowWithTheme";
+import renderWithTheme from "../../../../tests/hoc/shallowWithTheme";
 import Chip from "./Chip";
 
 it("renders article teaser correctly", () => {
