+++ src/components/Molecules/InfoBanner/InfoBanner.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import 'jest-styled-components';
-import renderWithTheme from '../../../hoc/shallowWithTheme';
+import renderWithTheme from '../../../../tests/hoc/shallowWithTheme';
 import InfoBanner from './InfoBanner';
 
 it('renders correctly', () => {
