+++ src/components/Molecules/SchoolLookup/SchoolLookup.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import 'jest-styled-components';
 
-import renderWithTheme from '../../../hoc/shallowWithTheme';
+import renderWithTheme from '../../../../tests/hoc/shallowWithTheme';
 import SchoolLookup from './SchoolLookup';
 
 it('renders correctly', () => {
