+++ src/components/Molecules/SearchInput/SearchInput.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import 'jest-styled-components';
-import renderWithTheme from '../../../hoc/shallowWithTheme';
+import renderWithTheme from '../../../../tests/hoc/shallowWithTheme';
 import SearchInput from './SearchInput';
 
 it('renders correctly', () => {
