+++ src/components/Molecules/ShareButton/ShareButton.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import 'jest-styled-components';
-import renderWithTheme from '../../../hoc/shallowWithTheme';
+import renderWithTheme from '../../../../tests/hoc/shallowWithTheme';
 import ShareButton from './ShareButton';
 
 it('renders correctly', () => {
