+++ src/components/Molecules/SingleMessage/SingleMessage.test.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import 'jest-styled-components';
-import renderWithTheme from '../../../hoc/shallowWithTheme';
+import renderWithTheme from '../../../../tests/hoc/shallowWithTheme';
 import SingleMessage from './SingleMessage';
 import { defaultData }  from '../../../styleguide/data/data';import Text from '../../Atoms/Text/Text';
 import Link from '../../Atoms/Link/Link';
