+++ src/components/Organisms/CookieBanner/CookieBanner.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "jest-styled-components";
-import renderWithTheme from "../../../hoc/shallowWithTheme";
+import renderWithTheme from "../../../../tests/hoc/shallowWithTheme";
 import CookieBanner from "./CookieBanner";
 
 it("renders correctly", () => {
