+++ src/components/Organisms/WYMDCarousel/WYMDCarousel.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import "jest-styled-components";
 
-import renderWithTheme from "../../../hoc/shallowWithTheme";
+import renderWithTheme from "../../../../tests/hoc/shallowWithTheme";
 import WYMDCarousel from "./WYMDCarousel";
 const { carouselItemsComplete } = require("../../../styleguide/data/data");
