+++ tests/hoc/shallowWithTheme.js
@@ -1,7 +1,8 @@
 import React from 'react';
 import renderer from 'react-test-renderer';
-import { crTheme } from '../index';
-import ThemeProvider from '../theme/ThemeProvider';
+
+import { crTheme } from '../../src/index';
+import ThemeProvider from '../../src/theme/ThemeProvider';
 
 export default function renderWithTheme(component) {
   return renderer.create(
